+++ app/src/sensors/l86_m33/zcbor/l86_m33_encode_types.h
@@ -26,7 +26,7 @@ extern "C" {
 #define DEFAULT_MAX_QTY 3
 
 struct L86_M33 {
-	int32_t timestamp;
+	int64_t timestamp;
 	double latitude;
 	double longitude;
 };
